Made it doubles

+++ Rms.java
@@ -6,15 +6,15 @@ public class Rms extends Thread {
     }
     public static void main(String[] args){
         //creation of 2d array for doWork function
-        int[][] arr = new int[10][10];
+        double[][] arr = new double[10][10];
         for(int i=0; i<arr.length; i++){
             for(int j=0; j<arr.length; j++){
-                arr[i][j]=1;
+                arr[i][j]=1.0;
             }
         }
         doWork(arr);
     }
-    public static void doWork(int [][] arr){
+    public static void doWork(double [][] arr){
         //Execution of doWork Function
         int counter=0;
         int place=0;
